pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "warp2api"
version = "0.1.0"
description = "Warp AI service to OpenAI API compatibility bridge"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]",
    "uvicorn[standard]",
    "httpx[http2]",
    "protobuf",
    "grpcio-tools",
    "python-dotenv",
    "websockets>=15.0.1",
    "openai>=1.106.0",
]

[project.scripts]
warp-server = "warp2api.app.bridge_runtime:main"
warp-openai = "warp2api.app.openai_runtime:main"
warp2api = "warp2api.app.main:main"
warp2api-bridge = "warp2api.app.bridge:main"
warp2api-openai = "warp2api.app.openai:main"
warp2api-proto = "warp2api.tools.proto_sync:main"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = [".", "src"]
include = ["warp2api*"]

[tool.setuptools.package-data]
warp2api = ["proto/*.proto"]

[[tool.uv.index]]
url = "https://mirrors.ustc.edu.cn/pypi/simple"
default = true

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
